+++ CHANGELOG.md
@@ -7,6 +7,8 @@ _Unreleased_
 * Amenity list is displayed top-down instead of left-right.
 * Added a draft version of the micromapping mode (can be enabled in Settings).
 * Added a draft version of the building & entrance editor (find it in Settings).
+* When your numeric keyboard cannot switch to letters, there's a fix in Settings.
+* If you prefer `contact:phone` and `contact:website`, there's a setting.
 * `PH off` support for opening hours and few more usability tweaks.
 * You can mark an unchanged amenity checked on the editor page.
 * Reduced distance for switching into the big map mode.